+++ CHANGES.rst
@@ -81,6 +81,7 @@ Documentation
   See :issue:`1017`.
 - Added Upgrade guide. See :issue:`997`.
 - Enable sphinx-issues extension. :issue:`1091`
+- Replaced "Arguments" and "Args" with "Parameters". :issue:`1076`
 - Added documentation of how to set up a development environment with git. :issue:`906`